+++ lib/src/widgets/phone_field_state.dart
@@ -1,14 +1,9 @@
 part of 'phone_field.dart';
 
-class _PhoneFieldState extends State<PhoneField> {
-  /// size of input so we can render inkwell at correct height
-  Size? _sizeInput;
-  Size? _countryCodeSize;
-
-  bool get _isOutlineBorder => widget.decoration.border is OutlineInputBorder;
+class PhoneFieldState extends State<PhoneField> {
   PhoneFieldController get controller => widget.controller;
 
-  _PhoneFieldState();
+  PhoneFieldState();
 
   @override
   void initState() {
@@ -27,6 +22,9 @@ class _PhoneFieldState extends State<PhoneField> {
   }
 
   void selectCountry() async {
+    if (!widget.isCountrySelectionEnabled) {
+      return;
+    }
     SystemChannels.textInput.invokeMethod('TextInput.hide');
     final selected = await widget.selectorNavigator.navigate(context);
     if (selected != null) {
@@ -38,45 +36,32 @@ class _PhoneFieldState extends State<PhoneField> {
 
   @override
   Widget build(BuildContext context) {
-    // The idea here is to have an InputDecorat with a prefix where the prefix
-    // is the flag + country code which visible (when focussed).
-    // Then we stack an InkWell with the country code (invisible) so
-    // it is the right width
+    // the idea here is to have a mouse region that surround every thing
+    // that has a text cursor.
+    // When the country chip is not shown it request focus.
+    // When the country chip is shown, clicking on it request country selection
     return MouseRegion(
       cursor: SystemMouseCursors.text,
-      child: Stack(
-        children: [
-          MeasureSize(
-            onChange: (size) => setState(() => _sizeInput = size),
-            child: GestureDetector(
-              onTap: () => controller.focusNode.requestFocus(),
-              child: _textField(),
-            ),
-          ),
-          if (controller.focusNode.hasFocus || controller.national != null)
-            _getInkWellOverlay(),
-        ],
-      ),
-    );
-  }
-
-  Widget _textField() {
-    return InputDecorator(
-      decoration: _getOutterInputDecoration(),
-      isFocused: controller.focusNode.hasFocus,
-      isEmpty: _isEffectivelyEmpty(),
-      child: Row(
-        children: [
-          Expanded(
+      child: GestureDetector(
+        onTap: controller.focusNode.requestFocus,
+        child: AbsorbPointer(
+          // absorb pointer when the country chip is not shown, else flutter
+          // still allows the country chip to be clicked even though it is not shown
+          absorbing: _isEffectivelyEmpty() && !controller.focusNode.hasFocus,
+          child: InputDecorator(
+            decoration: _getOutterInputDecoration(),
+            isFocused: controller.focusNode.hasFocus,
+            isEmpty: _isEffectivelyEmpty(),
             child: TextField(
               focusNode: controller.focusNode,
               controller: controller.nationalNumberController,
               enabled: widget.enabled,
               decoration: _getInnerInputDecoration(),
-              inputFormatters: [
-                FilteringTextInputFormatter.allow(RegExp(
-                    '[${Constants.plus}${Constants.digits}${Constants.punctuation}]')),
-              ],
+              inputFormatters: widget.inputFormatters ??
+                  [
+                    FilteringTextInputFormatter.allow(RegExp(
+                        '[${Constants.plus}${Constants.digits}${Constants.punctuation}]')),
+                  ],
               autofillHints: widget.autofillHints,
               keyboardType: widget.keyboardType,
               textInputAction: widget.textInputAction,
@@ -84,7 +69,6 @@ class _PhoneFieldState extends State<PhoneField> {
               strutStyle: widget.strutStyle,
               textAlign: widget.textAlign,
               textAlignVertical: widget.textAlignVertical,
-              textDirection: widget.textDirection,
               autofocus: widget.autofocus,
               obscuringCharacter: widget.obscuringCharacter,
               obscureText: widget.obscureText,
@@ -115,63 +99,34 @@ class _PhoneFieldState extends State<PhoneField> {
                   widget.enableIMEPersonalizedLearning,
             ),
           ),
-        ],
-      ),
-    );
-  }
-
-  /// gets the inkwell that is displayed on top of the input
-  /// for feedback on country code click
-  Widget _getInkWellOverlay() {
-    // width and height calculation are a bit hacky but a better way
-    // that works when the input is resized was not found
-    var height = _sizeInput?.height ?? 0;
-    var width = _countryCodeSize?.width ?? 0;
-    // when there is an error the widget height contains the error
-    // se we need to remove the error height
-    if (widget.errorText != null) {
-      height -= 24;
-    }
-
-    if (_isOutlineBorder) {
-      // outline border adds padding to the left
-      width += 12;
-    }
-
-    if (widget.decoration.prefixIconConstraints != null) {
-      width += widget.decoration.prefixIconConstraints!.maxWidth;
-    } else if (widget.decoration.prefixIcon != null) {
-      // prefix icon default size is 48px
-      width += 48;
-    }
-
-    return InkWell(
-      key: const ValueKey('country-code-overlay'),
-      onTap: () {},
-      onTapDown: (_) => selectCountry(),
-      child: SizedBox(
-        height: height,
-        width: width,
+        ),
       ),
     );
   }
 
   Widget _getCountryCodeChip() {
-    return MeasureSize(
-      onChange: (size) => setState(() => _countryCodeSize = size),
-      child: Padding(
-        padding: const EdgeInsets.fromLTRB(8, 0, 8, 0),
-        child: CountryCodeChip(
-          key: const ValueKey('country-code-chip'),
-          isoCode: controller.isoCode,
-          showFlag: widget.showFlagInInput,
-          textStyle: widget.countryCodeStyle ??
-              widget.decoration.labelStyle ??
-              TextStyle(
-                fontSize: 16,
-                color: Theme.of(context).textTheme.caption?.color,
-              ),
-          flagSize: widget.flagSize,
+    return MouseRegion(
+      cursor: SystemMouseCursors.click,
+      child: GestureDetector(
+        onTap: selectCountry,
+        // material here else the click pass through empty spaces
+        child: Material(
+          color: Colors.transparent,
+          child: Padding(
+            padding: const EdgeInsetsDirectional.fromSTEB(8, 0, 8, 0),
+            child: CountryCodeChip(
+              key: const ValueKey('country-code-chip'),
+              isoCode: controller.isoCode,
+              showFlag: widget.showFlagInInput,
+              textStyle: widget.countryCodeStyle ??
+                  widget.decoration.labelStyle ??
+                  TextStyle(
+                    fontSize: 16,
+                    color: Theme.of(context).textTheme.caption?.color,
+                  ),
+              flagSize: widget.flagSize,
+            ),
+          ),
         ),
       ),
     );
@@ -199,6 +154,7 @@ class _PhoneFieldState extends State<PhoneField> {
   }
 
   bool _isEffectivelyEmpty() {
+    if (widget.isCountryChipPersistent) return false;
     final outterDecoration = _getOutterInputDecoration();
     // when there is not label and an hint text we need to have
     // isEmpty false so the country code is displayed along the